Update BankLab
This update changes the menu system into switches. It is able to keep track of the last deposit and withdraw. Needs to be able to track last five transactions regardless of the category. Needs to have print out formatted as requested by instructor. Possibly need to add Paulas work as a function since i havent been able to figure out how to make a separate piece of code from another file run with this file.