Update BankLab
So right now I cannot get the system to work with the new filing system. So i reverted to the old file names for the meantime to finish up the remaining issues
with the file printing. math and stuff of that sort. The issue that i came accross when using the new filing system is that it exited out giving an error that it could not
find the file name. I created files for the accounting system and i do not know why it did not work. Like I said, i reverted some of the code back to the original so it runs
without errors for the most part. Paula, you probably understand the issue a bit better than i can at the moment. I added your new lines into the code. I tried to make it work
and so i moved on to continuing to make the rest of the code work. I figured the fileStructure function will be removed when we need to apply your stuff. Make the changes that you
see fit i guess. The printing stuff works so test it out before you make your changes and see what you think. Once you print into the receipt and the 10011.txt file to test out the
writing capabilities we have done so far, you get a better understanding what we have left. Going to work on my other tests i have to study for ~Nassar*/