Declare transaction type update
This update has the declaration of the type of transaction a user made in working order. So when someone uses the program, goes to deposit or withdraw, it will now declare its type when printed out in the receipt as well as show up in transaction history. I will be working on setting up the math resolving on the users balance to not include the balance. Because at the moment the user's balance is what ever the i[0] - i[5]. What i am going to change is that the first number is i[0] is the the balance the user is. The other numbers are just the transactions. Balance should be resolved prior to printing and not require it to be printed or showing in transaction history because it would be giving us the wrong number with the balance never changing if the user exits and tries again.